MGRAPPS-58 - extract dependencies validation in separate component, create mappers for dto
