+++ src/main/java/org/folio/am/controller/ApplicationController.java
@@ -10,7 +10,7 @@ import org.folio.am.domain.dto.ApplicationReferences;
 import org.folio.am.domain.dto.ValidationMode;
 import org.folio.am.domain.model.ValidationContext;
 import org.folio.am.rest.resource.ApplicationsApi;
-import org.folio.am.service.ApplicationInterfaceValidatorService;
+import org.folio.am.service.ApplicationReferencesValidationService;
 import org.folio.am.service.ApplicationService;
 import org.folio.am.service.ApplicationValidatorService;
 import org.springframework.http.ResponseEntity;
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ApplicationController extends BaseController implements ApplicationsApi {
 
   private final ApplicationValidatorService applicationValidatorService;
-  private final ApplicationInterfaceValidatorService applicationInterfaceValidatorService;
+  private final ApplicationReferencesValidationService applicationReferencesValidationService;
   private final ApplicationService applicationService;
 
   @Override
@@ -67,7 +67,7 @@ public class ApplicationController extends BaseController implements Application
 
   @Override
   public ResponseEntity<Void> validateModulesInterfaceIntegrity(ApplicationReferences applicationReferences) {
-    applicationInterfaceValidatorService.validate(applicationReferences);
+    applicationReferencesValidationService.validate(applicationReferences);
     return ResponseEntity.noContent().build();
   }
